migrate rpc/module tests to integration tests TODO:write unit tests
